+++ benches/examples_all_basic/distributed_calc.rs
@@ -103,8 +103,8 @@ fn endpoint_a(s: EndpointA) -> Result<(), Box<dyn Error>> {
 }
 
 fn endpoint_c(s: EndpointC<i32>) -> Result<(), Box<dyn Error>> {
-    let s = send_mpst_c_to_s(thread_rng().gen_range(1..=100) as i32, s);
-    let s = send_mpst_c_to_s(thread_rng().gen_range(1..=100) as i32, s);
+    let s = send_mpst_c_to_s(thread_rng().gen_range(1..=100), s);
+    let s = send_mpst_c_to_s(thread_rng().gen_range(1..=100), s);
 
     let choice: i32 = thread_rng().gen_range(1..=2);
